refactor: improved file namespacing
