+++ packages/developer-portal/src/components/apps/pipeline/pipeline-configure/github-repository-selection-element/index.ts
@@ -1,0 +1,1 @@
+export * from './github-repository-selection-element'
